Move cloud onboarding API to an onboarding platform
