+++ script/hassfest/dependencies.py
@@ -173,11 +173,10 @@ IGNORE_VIOLATIONS = {
     "logbook",
     # Temporary needed for migration until 2024.10
     ("conversation", "assist_pipeline"),
-    # The onboarding integration provides limited backup and cloud APIs for use
+    # The onboarding integration provides limited backup for use
     # during onboarding. The onboarding integration waits for the backup manager
-    # and cloud to be ready before calling any backup or cloud functionality.
+    # and to be ready before calling any backup functionality.
     ("onboarding", "backup"),
-    ("onboarding", "cloud"),
 }
